fix(codegen): render effect RHS per-backend from the typed AST (#143 #144 #145 #146)
xeroc's v2.38.0 Kani-codegen quartet shared one root cause: the
adapter's render_effect lowered compound effect RHS via expr_to_lean
with an empty type env and no canonicalization, and that single
Lean-form string flowed verbatim into MIR Expr.rust (from_raw) and out
through every backend. The Lean backend was bitten too — `residual :=
fee - cut` rendered as `s.fee - cut` (heuristic front-prefix only),
leaving `cut` unbound.

Fixes, per issue:

- #143/#144 — render_effect_rhs_forms canonicalizes the RHS AST
  (canon.rs seam from #139) and renders BOTH forms: Lean via
  expr_to_lean with the real env, Rust via expr_to_rust. The Rust form
  threads through ParsedHandler::effects_rust / ParsedEffectArm ->
  lower_effects -> MIR Expr.rust, so the Kani/proptest transition
  bodies and conformance asserts receive `bps_mul(amount, s.rate)` and
  `if s.flag == 1 { … } else { 0 }` instead of ML application syntax.
  Simple RHS shapes keep the legacy single-form string, so downstream
  binder-specific resolution (resolve_value, Anchor mechanize_effect)
  is byte-identical.

- #145 — guards_use_math_helpers now probes ref_impl bodies and effect
  RHS, so `mul_div_floor` used only inside a ref_impl emits the
  mul_div_floor_u128 definition; the ref_impl body narrows back to its
  declared return width (`(…) as u64`), same precedent as the
  let-binding narrowing.

- #146 — two-sided lowering:
    * Predicates (requires / properties / ensures): new math-exact
      render (RustOpts::widen_arith) stored in rust_expr_math /
      rust_expression_math / rust_expr_binary_math and preferred by the
      harness emitters. Arithmetic inside comparisons evaluates in
      u128/i128 (`-` on unsigned kinds saturates = Nat monus, `/`//`%`
      follow Lean's total-function convention), exactly matching the
      Lean Nat model — no more spurious Kani overflow failures on
      unconstrained symbolic state. The `(a * b) / 10000` shape is
      exempt so the solver-tuned mul_bps_floor_u128 rewrite keeps
      firing. Scaffold-facing rust_expr is untouched.
    * Effect RHS: bare arithmetic lowers checked
      (RustOpts::checked_arith → `checked_sub(..)?` inside an
      `(|| Some(…))()` closure, None → return false), extending the
      v2.7 checked `+=`/`-=` doctrine to `:=` bodies.

Known divergence, documented in the DSL reference and the fixture
README: Lean Nat subtraction in a `:=` RHS is monus while the harness
rejects on underflow; aligning Lean via auto bound-guards is follow-up.

Regression fixture: tests/fixtures/regressions/issues-143-146-kani-arith
(vault.qedspec distilled from all four reports), pinned by
kani_mir::tests::compound_effect_rhs_and_arith_predicates_render_soundly
(incl. a syn parse gate on the whole harness) and
lean_gen_mir::tests::compound_effect_rhs_lean_is_fully_state_qualified.
Snapshots + bundled multisig/percolator harnesses regenerated
(qedgen check --regen-drift --write); Anchor/Quasar scaffold outputs are
byte-identical (codegen_snapshot untouched).

+++ crates/qedgen/tests/snapshots/multisig.proptest.rs
@@ -91,7 +91,7 @@ fn threshold_bounded(s: &State) -> bool {
 
 /// votes_bounded: s.approval_count + s.rejection_count ≤ s.member_count
 fn votes_bounded(s: &State) -> bool {
-    s.approval_count + s.rejection_count <= s.member_count
+    ((s.approval_count) as u128) + ((s.rejection_count) as u128) <= ((s.member_count) as u128)
 }
 
 fn create_vault(s: &mut State, threshold: u8, member_count: u8) -> bool {
@@ -159,7 +159,7 @@ fn execute(s: &mut State, member_index: u8) -> bool {
 }
 
 fn cancel_proposal(s: &mut State) -> bool {
-    if !((s.member_count.saturating_sub(s.rejection_count) < s.threshold)) {
+    if !(((((s.member_count) as u128)).saturating_sub(((s.rejection_count) as u128)) < ((s.threshold) as u128))) {
         return false;
     }
     if s.status != Status::HasProposal {
@@ -454,7 +454,7 @@ proptest! {
     #[test]
     fn cancel_proposal_rejects_invalid(s in arb_boundary_state()) {
         let mut s = s;
-        prop_assume!(!((s.member_count.saturating_sub(s.rejection_count) < s.threshold)));
+        prop_assume!(!(((((s.member_count) as u128)).saturating_sub(((s.rejection_count) as u128)) < ((s.threshold) as u128))));
         prop_assert!(!cancel_proposal(&mut s),
             "cancel_proposal must reject when guard is violated");
     }
