+++ crates/qedgen/tests/snapshots/percolator.kani.rs
@@ -80,7 +80,7 @@ struct State {
 
 /// conservation: s.V ≥ ((∑ i : AccountIdx, s.accounts[i].capital)) + ((∑ i : AccountIdx, s.accounts[i].reserved_pnl)) + s.I + s.F
 fn conservation(s: &State) -> bool {
-    s.V >= (sum_over::<AccountIdx>(|i| s.accounts[(i) as usize].capital)) + (sum_over::<AccountIdx>(|i| s.accounts[(i) as usize].reserved_pnl)) + s.I + s.F
+    ((s.V) as u128) >= (((sum_over::<AccountIdx>(|i| s.accounts[(i) as usize].capital)) as u128)) + (((sum_over::<AccountIdx>(|i| s.accounts[(i) as usize].reserved_pnl)) as u128)) + ((s.I) as u128) + ((s.F) as u128)
 }
 
 /// vault_bounded: s.V ≤ 10000000000000000
@@ -108,7 +108,7 @@ fn account_solvent_at(s: &State, i: usize) -> bool {
 // ============================================================================
 
 fn add_user(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) {
+    if !((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -120,7 +120,7 @@ fn add_user(s: &mut State, i: usize) -> bool {
 }
 
 fn add_lp(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) {
+    if !((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -161,7 +161,7 @@ fn close_account(s: &mut State, i: usize) -> bool {
 }
 
 fn deposit(s: &mut State, i: usize, amount: u128) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (s.V + amount <= 10000000000000000)) {
+    if !((s.accounts[(i) as usize].active == 1) && (((s.V) as u128) + ((amount) as u128) <= ((10000000000000000) as u128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -180,7 +180,7 @@ fn deposit(s: &mut State, i: usize, amount: u128) -> bool {
 }
 
 fn withdraw(s: &mut State, i: usize, amount: u128) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((amount) as i128))) {
+    if !((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((amount) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -199,7 +199,7 @@ fn withdraw(s: &mut State, i: usize, amount: u128) -> bool {
 }
 
 fn top_up_insurance(s: &mut State, amount: u128) -> bool {
-    if !((s.V + amount <= 10000000000000000)) {
+    if !((((s.V) as u128) + ((amount) as u128) <= ((10000000000000000) as u128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -218,7 +218,7 @@ fn top_up_insurance(s: &mut State, amount: u128) -> bool {
 }
 
 fn deposit_fee_credits(s: &mut State, i: usize, amount: u128) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (s.V + amount <= 10000000000000000)) {
+    if !((s.accounts[(i) as usize].active == 1) && (((s.V) as u128) + ((amount) as u128) <= ((10000000000000000) as u128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -271,7 +271,7 @@ fn execute_trade(s: &mut State, a: usize, b: usize, size_q: i128, exec_price: u6
 }
 
 fn liquidate_case_0(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128)) && (false)) {
+    if !((s.accounts[(i) as usize].active == 1) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128)) && (false)) {
         return false;
     }
     if s.status != Status::Active {
@@ -282,7 +282,7 @@ fn liquidate_case_0(s: &mut State, i: usize) -> bool {
 }
 
 fn liquidate_case_1(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))) {
+    if !((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) + ((s.I) as i128) >= ((0) as i128))) {
         return false;
     }
     if s.status != Status::Active {
@@ -294,7 +294,7 @@ fn liquidate_case_1(s: &mut State, i: usize) -> bool {
 }
 
 fn liquidate_otherwise(s: &mut State, i: usize) -> bool {
-    if !((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (!(((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))) && (false)) {
+    if !((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))) && (!(((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) + ((s.I) as i128) >= ((0) as i128))) && (false)) {
         return false;
     }
     if s.status != Status::Active {
@@ -357,7 +357,7 @@ fn verify_add_user_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!add_user(&mut s, i),
         "add_user must reject when guard is violated");
@@ -377,7 +377,7 @@ fn verify_add_lp_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 0) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!add_lp(&mut s, i),
         "add_lp must reject when guard is violated");
@@ -438,7 +438,7 @@ fn verify_deposit_rejects_invalid() {
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
     let amount: u128 = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (s.V + amount <= 10000000000000000)));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (((s.V) as u128) + ((amount) as u128) <= ((10000000000000000) as u128))));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!deposit(&mut s, i, amount),
         "deposit must reject when guard is violated");
@@ -459,7 +459,7 @@ fn verify_withdraw_rejects_invalid() {
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
     let amount: u128 = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((amount) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (s.accounts[(i) as usize].capital >= amount) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((amount) as i128))));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!withdraw(&mut s, i, amount),
         "withdraw must reject when guard is violated");
@@ -479,7 +479,7 @@ fn verify_top_up_insurance_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let amount: u128 = kani::any();
-    kani::assume(!((s.V + amount <= 10000000000000000)));
+    kani::assume(!((((s.V) as u128) + ((amount) as u128) <= ((10000000000000000) as u128))));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!top_up_insurance(&mut s, amount),
         "top_up_insurance must reject when guard is violated");
@@ -500,7 +500,7 @@ fn verify_deposit_fee_credits_rejects_invalid() {
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
     let amount: u128 = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (s.V + amount <= 10000000000000000)));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (((s.V) as u128) + ((amount) as u128) <= ((10000000000000000) as u128))));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!deposit_fee_credits(&mut s, i, amount),
         "deposit_fee_credits must reject when guard is violated");
@@ -564,7 +564,7 @@ fn verify_liquidate_case_0_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128)) && (false)));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128)) && (false)));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!liquidate_case_0(&mut s, i),
         "liquidate_case_0 must reject when guard is violated");
@@ -584,7 +584,7 @@ fn verify_liquidate_case_1_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))) && (((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) + ((s.I) as i128) >= ((0) as i128))));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!liquidate_case_1(&mut s, i),
         "liquidate_case_1 must reject when guard is violated");
@@ -604,7 +604,7 @@ fn verify_liquidate_otherwise_rejects_invalid() {
     };
     kani::assume(s.status == Status::Active);
     let i: usize = kani::any();
-    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))) && (!(((((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) + ((s.I) as i128)) as i128) >= ((0) as i128))) && (false)));
+    kani::assume(!((s.accounts[(i) as usize].active == 1) && (!(((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) >= ((0) as i128))) && (!(((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128) + ((s.I) as i128) >= ((0) as i128))) && (false)));
     kani::cover!(true, "guard-violation domain is satisfiable");
     assert!(!liquidate_otherwise(&mut s, i),
         "liquidate_otherwise must reject when guard is violated");
